refactor: png 파일 svg파일로 교체
